ft: plan Management

+++ src/Model/Plan.java
@@ -12,15 +12,15 @@ public class Plan {
     private int plan_id;
     private String name;
     private long price;
-    private int max_speed_domestic;
-    private int max_speed_international;
-    private int min_speed_domestic;
-    private int min_speed_international;
+    private double max_speed_domestic;
+    private double max_speed_international;
+    private double min_speed_domestic;
+    private double min_speed_international;
 
     public Plan() {
     }
 
-    public Plan(int plan_id, String name, long price, int max_speed_domestic, int max_speed_international, int min_speed_domestic, int min_speed_international) {
+    public Plan(int plan_id, String name, long price, double max_speed_domestic, double max_speed_international, double min_speed_domestic, double min_speed_international) {
         this.plan_id = plan_id;
         this.name = name;
         this.price = price;
@@ -54,35 +54,35 @@ public class Plan {
         this.price = price;
     }
 
-    public int getMax_speed_domestic() {
+    public double getMax_speed_domestic() {
         return max_speed_domestic;
     }
 
-    public void setMax_speed_domestic(int max_speed_domestic) {
+    public void setMax_speed_domestic(double max_speed_domestic) {
         this.max_speed_domestic = max_speed_domestic;
     }
 
-    public int getMax_speed_international() {
+    public double getMax_speed_international() {
         return max_speed_international;
     }
 
-    public void setMax_speed_international(int max_speed_international) {
+    public void setMax_speed_international(double max_speed_international) {
         this.max_speed_international = max_speed_international;
     }
 
-    public int getMin_speed_domestic() {
+    public double getMin_speed_domestic() {
         return min_speed_domestic;
     }
 
-    public void setMin_speed_domestic(int min_speed_domestic) {
+    public void setMin_speed_domestic(double min_speed_domestic) {
         this.min_speed_domestic = min_speed_domestic;
     }
 
-    public int getMin_speed_international() {
+    public double getMin_speed_international() {
         return min_speed_international;
     }
 
-    public void setMin_speed_international(int min_speed_international) {
+    public void setMin_speed_international(double min_speed_international) {
         this.min_speed_international = min_speed_international;
     }
